Small fix to thread allocation for NetWare
Obtained from:  new-httpd
Submitted by:	Brad Nicholes <BNICHOLES@novell.com>
Reviewed by:	William Rowe


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@86908 13f79535-47bb-0310-9956-ffa450edef68

+++ src/os/netware/os.h
@@ -67,7 +67,7 @@
  * and prototypes of OS specific functions defined in os.c or os-inline.c
  */
 
-typedef int int32;
+typedef signed long int32;
 typedef int uid_t;
 typedef int gid_t;
